[feat] basic ops done

+++ src/services/task.ts
@@ -1,11 +1,30 @@
 import { taskRepository } from "../repositories/task.ts";
-import { taskSchema, type CreateTask, type Task, type UpdateTask } from "../schemas/task.ts";
+import { taskSchema, type BaseTaskIdentifiers, type CreateTask, type RenameTask, type Task, type UpdateDescriptionTask, type UpdateLabelsTask, type UpdateMembersTask } from "../schemas/task.ts";
 import { InternalServerErrorException } from "../utils/exceptions/server.ts";
+import { v4 as uuid } from "uuid";
+import { workSpaceServices } from "./workspace.ts";
+import { BadRequestException, UnauthorizedException } from "../utils/exceptions/client.ts";
 
 class TaskServices {
-    async createTask({title, description, members, boardId, listId, workspaceId, createdBy}: CreateTask): Promise<Task> {
+    async createTask({ title, description, members, boardId, boardListId, workspaceId, createdBy, labels }: CreateTask): Promise<Task> {
+        const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+        if(!isEligible){
+            throw new UnauthorizedException("You are not authorized to create a task");
+        }
         try {
-            const task = await taskRepository.createTask({title, description, members, boardId, listId, workspaceId, createdBy});
+            const newTask = {
+                taskId: `t-${uuid()}`,
+                title,
+                description,
+                members,
+                boardId,
+                boardListId,
+                workspaceId,
+                createdBy,
+                labels,
+                createdAt: new Date().toISOString(),
+            }
+            const task = await taskRepository.createTask(newTask);
             return taskSchema.parse(task);
         } catch (error) {
             throw new InternalServerErrorException("Internal Server Error");
@@ -27,15 +46,18 @@ class TaskServices {
             if(!task){
                 return null;
             }
-            const { _id, ...rest } = task;
-            return task ? taskSchema.parse(rest) : null;
+            return task ? taskSchema.parse(task) : null;
         } catch (error) {
             throw new InternalServerErrorException("Internal Server Error");
         }
     }
 
-    async deleteTask(taskId: string): Promise<boolean> {
+    async deleteTask(taskId: string, workspaceId: string, createdBy: string): Promise<boolean> {
         try {
+            const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+            if(!isEligible){
+                throw new UnauthorizedException("You are not authorized to delete this task");
+            }
             const deleted = await taskRepository.deleteTask(taskId);
             return deleted;
         } catch (error) {
@@ -43,16 +65,78 @@ class TaskServices {
         }
     }
 
-    async updateTask({ taskId, title, description, members, boardId, listId, workspaceId }: UpdateTask): Promise<Task | null> {
+    async renameTask({taskId, workspaceId, boardId, boardListId, title, createdBy}:RenameTask):Promise<void>{
+        if(!createdBy){
+            throw new BadRequestException("missing required fields");
+        }
         try {
-            const task = await taskRepository.updateTask({ taskId, title, description, members, boardId, listId, workspaceId });
-            if(!task){
-                return null;
+            const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+            if(!isEligible){
+                throw new UnauthorizedException("You are not authorized to rename this task");
             }
-            const { _id, ...rest } = task;
-            return task ? taskSchema.parse(rest) : null;
+            await taskRepository.renameTask({taskId, workspaceId, boardId, boardListId, title});
         } catch (error) {
-            throw new InternalServerErrorException("Internal Server Error");
+            throw new InternalServerErrorException(String(error));
+        }
+    }
+
+    async updateDescription({taskId, workspaceId, boardId, boardListId, description, createdBy}: UpdateDescriptionTask): Promise<void>{
+        try {
+           if(!createdBy){
+                throw new BadRequestException("missing required fields");
+            }
+            const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+            if(!isEligible){
+                throw new UnauthorizedException("You are not authorized to update this task");
+            }
+            await taskRepository.updateDescription({taskId, workspaceId, boardId, boardListId, description});
+        } catch (error) {
+            throw new InternalServerErrorException(String(error));
+        }
+    }
+
+    async updateMembers({taskId, workspaceId, boardId, boardListId, members, createdBy}: UpdateMembersTask): Promise<void>{
+        try {
+            if(!createdBy){
+                throw new BadRequestException("missing required fields");
+            }
+            const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+            if(!isEligible){
+                throw new UnauthorizedException("You are not authorized to update this task");
+            }
+            await taskRepository.updateMembers({taskId, workspaceId, boardId, boardListId, members});
+        } catch (error) {
+            throw new InternalServerErrorException(String(error));
+        }
+    }
+
+    async updateListId({ taskId, workspaceId, boardId, boardListId, createdBy }: BaseTaskIdentifiers): Promise<void>{
+        try {
+            if(!createdBy){
+                throw new BadRequestException("missing required fields");
+            }
+            const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+            if(!isEligible){
+                throw new UnauthorizedException("You are not authorized to update this task");
+            }
+            await taskRepository.updateListId({ taskId, workspaceId, boardId, boardListId });
+        } catch (error) {
+            throw new InternalServerErrorException(String(error));
+        }
+    }
+
+    async updateLabels({ taskId, workspaceId, boardId, boardListId, labels, createdBy }: UpdateLabelsTask): Promise<void>{
+        try {
+            if(!createdBy){
+                throw new BadRequestException("missing required fields");
+            }
+            const isEligible = await workSpaceServices.authorizeMembers(workspaceId, createdBy);
+            if(!isEligible){
+                throw new UnauthorizedException("You are not authorized to update this task");
+            }
+            await taskRepository.updateLabels({ taskId, workspaceId, boardId, boardListId, labels });
+        } catch (error) {
+            throw new InternalServerErrorException(String(error));
         }
     }
 }
